refactored and improved code for arithmetic evaluator

+++ arithmetic-evaluator/include/project/posftix_evaluation.hpp
@@ -1,5 +1,4 @@
 #include "project/notation_conversion.hpp"
 #include <vector>
 
-
-double eval_postfix(vector<token> postfix_tokens);
+float eval_postfix(vector<token> postfix_tokens);
